Broadcast the dbouble spend from within the respendrelayer trigger
